move schema adaptation into data.adapt
