+++ lenskit/lenskit/data/__init__.py
@@ -9,9 +9,9 @@ Data abstractions and data set access.
 
 from __future__ import annotations
 
+from .adapt import from_interactions_df
 from .builder import DatasetBuilder
 from .collection import GenericKey, ItemListCollection, UserIDKey
-from .convert import from_interactions_df
 from .dataset import Dataset, FieldError
 from .items import ItemList
 from .lazy import LazyDataset
