+++ lenskit/tests/data/test_adapt_schemas.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 from pytest import raises, warns
 
-from lenskit.data.schemas import normalize_columns
+from lenskit.data.adapt import normalize_columns
 from lenskit.data.types import AliasedColumn
 
 SCROLLS = pd.DataFrame(
