+++ lenskit/tests/data/test_dataset_convert.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from lenskit.data.convert import from_interactions_df
+from lenskit.data.adapt import from_interactions_df
 
 
 def test_item_subset(rng: np.random.Generator, ml_ratings: pd.DataFrame):
